% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funLevelsDS.R
\name{.checkLevelsDisclosureRisk}
\alias{.checkLevelsDisclosureRisk}
\title{Check Disclosure Risk for Levels}
\usage{
.checkLevelsDisclosureRisk(input, levels_out)
}
\arguments{
\item{input}{A vector of values from which levels are derived.}

\item{levels_out}{A vector containing the levels to be checked.}
}
\value{
No return value; throws an error if disclosure risk is too high.
}
\description{
This function checks whether the levels of a variable can be safely returned
based on the disclosure risk settings.
}
\keyword{internal}
