% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funLevelsDS.R
\name{.getDensitySetting}
\alias{.getDensitySetting}
\title{Get Disclosure Density Setting}
\usage{
.getDensitySetting()
}
\value{
A numeric value representing the density setting.
}
\description{
Retrieves the \code{nfilter.levels.density} setting from DataSHIELD's disclosure settings.
}
\keyword{internal}
