% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funLevelsDS.R
\name{.throwErrorIfRisk}
\alias{.throwErrorIfRisk}
\title{Throw an Error if Disclosure Risk is Too High}
\usage{
.throwErrorIfRisk(input, levels_out, threshold)
}
\arguments{
\item{input}{A vector of values.}

\item{levels_out}{A vector of levels being checked.}

\item{threshold}{A numeric value representing the maximum allowed levels.}
}
\value{
No return value; throws an error if the disclosure risk is exceeded.
}
\description{
Internal function that throws an error if the number of levels exceeds the
threshold set by disclosure risk settings.
}
\keyword{internal}
