% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funLevelsDS.R
\name{funLevelsDS}
\alias{funLevelsDS}
\title{Returns the levels of a factor vector with a fun message}
\usage{
funLevelsDS(x, fun_message)
}
\arguments{
\item{x}{a factor vector}

\item{fun_message}{a fun message}
}
\value{
a list, the factor levels present in the vector
}
\description{
Returns the levels of a factor vector with a fun message
}
\author{
Tim Cadman
}
